refactor: improve code structure, create classes for data entities
