+++ app/src/main/java/com/effectivelab1/heroapp/data/database/AppDatabase.kt
@@ -1,9 +1,10 @@
-package com.effectivelab1.heroapp.data
+package com.effectivelab1.heroapp.data.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.effectivelab1.heroapp.data.model.MarvelCharacterEntity
 
 @Database(entities = [MarvelCharacterEntity::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
