+++ app/src/main/java/com/effectivelab1/heroapp/data/database/CharacterDao.kt
@@ -1,10 +1,11 @@
 
-package com.effectivelab1.heroapp.data
+package com.effectivelab1.heroapp.data.database
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.effectivelab1.heroapp.data.model.MarvelCharacterEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
