+++ app/src/main/java/com/effectivelab1/heroapp/data/model/CharacterEntity.kt
@@ -1,4 +1,4 @@
-package com.effectivelab1.heroapp.data
+package com.effectivelab1.heroapp.data.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
