+++ app/src/main/java/com/effectivelab1/heroapp/data/model/CharacterUI.kt
@@ -1,0 +1,8 @@
+package com.effectivelab1.heroapp.data.model
+
+data class MarvelCharacterUI(
+    val id: Int,
+    val name: String,
+    val description: String,
+    val imageUrl: String
+)
