+++ app/src/main/java/com/effectivelab1/heroapp/presentation/screens/HeroDetailScreen.kt
@@ -1,35 +1,28 @@
-package com.effectivelab1.heroapp.presentation.screens.heroInfoScreen
+package com.effectivelab1.heroapp.presentation.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shadow
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.navigation.NavController
-import com.effectivelab1.heroapp.R
-import com.effectivelab1.heroapp.constants.Constants
-import com.effectivelab1.heroapp.constants.Constants.iconButtonPaddingStart
-import com.effectivelab1.heroapp.constants.Constants.sizeIconArrowBack
-import com.effectivelab1.heroapp.presentation.components.ImageLoader
-import com.effectivelab1.heroapp.presentation.models.MarvelCharacter
+import com.effectivelab1.heroapp.util.Constants
+import com.effectivelab1.heroapp.presentation.screens.components.ImageLoader
+import com.effectivelab1.heroapp.data.model.MarvelCharacter
+import com.effectivelab1.heroapp.data.model.MarvelCharacterUI
+import com.effectivelab1.heroapp.presentation.screens.components.LoadingText
 import com.effectivelab1.heroapp.presentation.viewModel.CharacterViewModel
+import com.effectivelab1.heroapp.presentation.screens.components.NavigationBackButton
 
 @Composable
 fun HeroDetailScreen(
@@ -45,11 +38,7 @@ fun HeroDetailScreen(
     if (currentHero == null) {
         Box(modifier = Modifier.fillMaxSize()) {
             NavigationBackButton(navigator = navigator)
-            Text(
-                text = stringResource(R.string.loading_string),
-                color = Color.Black,
-                modifier = Modifier.align(Alignment.Center),
-            )
+            LoadingText()
         }
     } else {
         Box(modifier = Modifier.fillMaxSize()) {
@@ -65,7 +54,7 @@ fun HeroDetailScreen(
 }
 
 @Composable
-private fun HeroInformation(currentHero: MarvelCharacter) {
+private fun HeroInformation(currentHero: MarvelCharacterUI) {
     Column(
         modifier =
         Modifier
@@ -121,22 +110,4 @@ private fun HeroDescription(description: String) {
     )
 }
 
-@Composable
-private fun NavigationBackButton(
-    navigator: NavController,
-    modifier: Modifier = Modifier,
-) {
-    IconButton(
-        onClick = { navigator.popBackStack() },
-        modifier =
-        modifier
-            .padding(iconButtonPaddingStart)
-            .size(sizeIconArrowBack),
-    ) {
-        Icon(
-            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-            contentDescription = "Back",
-            tint = Color.LightGray,
-        )
-    }
-}
+
