+++ app/src/main/java/com/effectivelab1/heroapp/util/CharacterMapper.kt
@@ -1,25 +1,30 @@
 package com.effectivelab1.heroapp.util
 
-import com.effectivelab1.heroapp.data.MarvelCharacterEntity
-import com.effectivelab1.heroapp.presentation.models.MarvelCharacter
-import com.effectivelab1.heroapp.presentation.models.MarvelImage
+import com.effectivelab1.heroapp.data.model.MarvelCharacterEntity
+import com.effectivelab1.heroapp.data.model.MarvelCharacter
+import com.effectivelab1.heroapp.data.model.MarvelCharacterUI
+import com.effectivelab1.heroapp.data.model.MarvelImage
 
 fun MarvelCharacter.toEntity(): MarvelCharacterEntity =
     MarvelCharacterEntity(
         id = this.id,
         name = this.name,
         description = this.description,
-        imageUrl = this.imageUrl,
+        imageUrl = this.imageUrl
     )
 
-fun MarvelCharacterEntity.toUI(): MarvelCharacter =
-    MarvelCharacter(
+fun MarvelCharacterEntity.toUI(): MarvelCharacterUI =
+    MarvelCharacterUI(
         id = this.id,
         name = this.name,
         description = this.description,
-        thumbnail =
-        MarvelImage(
-            path = this.imageUrl.substringBeforeLast('.'),
-            extension = this.imageUrl.substringAfterLast('.'),
-        ),
+        imageUrl = this.imageUrl
+    )
+
+fun MarvelCharacter.toUI(): MarvelCharacterUI =
+    MarvelCharacterUI(
+        id = this.id,
+        name = this.name,
+        description = this.description,
+        imageUrl = this.imageUrl
     )
